LPD-75170 refactor: rename ai-hub-rest-impl to ai-hub-nexus-rest-impl
