+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-rest-impl/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay AI Hub Nexus REST Implementation
+Bundle-SymbolicName: com.liferay.ai.hub.nexus.rest.impl
+Bundle-Version: 1.0.6
