+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-rest-impl/src/main/java/com/liferay/ai/hub/nexus/rest/internal/resource/v1_0/ChatResourceImpl.java
@@ -1,0 +1,42 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.ai.hub.nexus.rest.internal.resource.v1_0;
+
+import com.liferay.ai.hub.nexus.rest.resource.v1_0.ChatResource;
+import com.liferay.ai.hub.nexus.rest.resource.v1_0.util.SseUtil;
+import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
+
+import jakarta.ws.rs.core.Context;
+import jakarta.ws.rs.sse.Sse;
+import jakarta.ws.rs.sse.SseEventSink;
+
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.ServiceScope;
+
+/**
+ * @author Feliphe Marinho
+ */
+@Component(
+	properties = "OSGI-INF/liferay/rest/v1_0/chat.properties",
+	scope = ServiceScope.PROTOTYPE, service = ChatResource.class
+)
+public class ChatResourceImpl extends BaseChatResourceImpl {
+
+	@Override
+	public void getChatSubscribe(SseEventSink sseEventSink) {
+		if (!FeatureFlagManagerUtil.isEnabled(
+				contextCompany.getCompanyId(), "LPD-62272")) {
+
+			throw new UnsupportedOperationException();
+		}
+
+		SseUtil.initialize(_sse, sseEventSink);
+	}
+
+	@Context
+	private Sse _sse;
+
+}
