+++ modules/dxp/apps/ai-hub-nexus/ai-hub-nexus-rest-impl/src/main/resources/OSGI-INF/liferay/rest/v1_0/agent-definition.properties
