--- modules/dxp/apps/ai-hub-nexus/ai-hub-rest-impl/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay AI Hub REST Implementation
-Bundle-SymbolicName: com.liferay.ai.hub.rest.impl
-Bundle-Version: 1.0.6
