--- modules/dxp/apps/ai-hub-nexus/ai-hub-rest-impl/rest-config.yaml
@@ -1,13 +1,0 @@
-apiDir: ../ai-hub-rest-api/src/main/java
-apiPackagePath: com.liferay.ai.hub.rest
-application:
-    baseURI: /ai-hub
-    className: AIHubRESTApplication
-    name: Liferay.AI.Hub.REST
-author: Feliphe Marinho
-clientDir: ../ai-hub-rest-client/src/main/java
-compatibilityVersion: 13
-forcePredictableOperationId: true
-generateGraphQL: false
-javaEEPackage: jakarta
-testDir: ../ai-hub-rest-test/src/testIntegration/java
