--- modules/dxp/apps/ai-hub-nexus/ai-hub-rest-impl/src/main/java/com/liferay/ai/hub/rest/internal/jaxrs/application/AIHubRESTApplication.java
@@ -1,28 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.hub.rest.internal.jaxrs.application;
-
-import jakarta.annotation.Generated;
-
-import jakarta.ws.rs.core.Application;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Feliphe Marinho
- * @generated
- */
-@Component(
-	property = {
-		"liferay.jackson=false", "osgi.jaxrs.application.base=/ai-hub",
-		"osgi.jaxrs.extension.select=(osgi.jaxrs.name=Liferay.Vulcan)",
-		"osgi.jaxrs.name=Liferay.AI.Hub.REST"
-	},
-	service = Application.class
-)
-@Generated("")
-public class AIHubRESTApplication extends Application {
-}
