--- modules/dxp/apps/ai-hub-nexus/ai-hub-rest-impl/src/main/java/com/liferay/ai/hub/rest/internal/resource/v1_0/BaseChatResourceImpl.java
@@ -1,505 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.ai.hub.rest.internal.resource.v1_0;
-
-import com.liferay.ai.hub.nexus.rest.resource.v1_0.ChatResource;
-import com.liferay.petra.function.UnsafeFunction;
-import com.liferay.petra.function.transform.TransformUtil;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.security.permission.resource.ModelResourcePermission;
-import com.liferay.portal.kernel.service.GroupLocalService;
-import com.liferay.portal.kernel.service.ResourceActionLocalService;
-import com.liferay.portal.kernel.service.ResourcePermissionLocalService;
-import com.liferay.portal.kernel.service.RoleLocalService;
-import com.liferay.portal.odata.filter.ExpressionConvert;
-import com.liferay.portal.odata.filter.FilterParserProvider;
-import com.liferay.portal.odata.sort.SortParserProvider;
-import com.liferay.portal.vulcan.accept.language.AcceptLanguage;
-import com.liferay.portal.vulcan.util.ActionUtil;
-import com.liferay.portal.vulcan.util.UriInfoUtil;
-
-import jakarta.annotation.Generated;
-
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import jakarta.ws.rs.core.UriInfo;
-
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-
-/**
- * @author Feliphe Marinho
- * @generated
- */
-@Generated("")
-@jakarta.ws.rs.Path("/v1.0")
-public abstract class BaseChatResourceImpl implements ChatResource {
-
-	/**
-	 * Invoke this method with the command line:
-	 *
-	 * curl -X 'GET' 'http://localhost:8080/o/ai-hub/v1.0/chats/subscribe'  -u 'test@liferay.com:test'
-	 */
-	@io.swagger.v3.oas.annotations.tags.Tags(
-		value = {@io.swagger.v3.oas.annotations.tags.Tag(name = "Chat")}
-	)
-	@jakarta.ws.rs.GET
-	@jakarta.ws.rs.Path("/chats/subscribe")
-	@jakarta.ws.rs.Produces("text/event-stream")
-	@Override
-	public void getChatSubscribe(
-			@jakarta.ws.rs.core.Context jakarta.ws.rs.sse.SseEventSink
-				sseEventSink)
-		throws Exception {
-	}
-
-	public void setContextAcceptLanguage(AcceptLanguage contextAcceptLanguage) {
-		this.contextAcceptLanguage = contextAcceptLanguage;
-	}
-
-	public void setContextCompany(
-		com.liferay.portal.kernel.model.Company contextCompany) {
-
-		this.contextCompany = contextCompany;
-	}
-
-	public void setContextHttpServletRequest(
-		HttpServletRequest contextHttpServletRequest) {
-
-		this.contextHttpServletRequest = contextHttpServletRequest;
-	}
-
-	public void setContextHttpServletResponse(
-		HttpServletResponse contextHttpServletResponse) {
-
-		this.contextHttpServletResponse = contextHttpServletResponse;
-	}
-
-	public void setContextUriInfo(UriInfo contextUriInfo) {
-		this.contextUriInfo = UriInfoUtil.getVulcanUriInfo(
-			getApplicationPath(), contextUriInfo);
-	}
-
-	public void setContextUser(
-		com.liferay.portal.kernel.model.User contextUser) {
-
-		this.contextUser = contextUser;
-	}
-
-	public void setExpressionConvert(
-		ExpressionConvert<com.liferay.portal.kernel.search.filter.Filter>
-			expressionConvert) {
-
-		this.expressionConvert = expressionConvert;
-	}
-
-	public void setFilterParserProvider(
-		FilterParserProvider filterParserProvider) {
-
-		this.filterParserProvider = filterParserProvider;
-	}
-
-	public void setGroupLocalService(GroupLocalService groupLocalService) {
-		this.groupLocalService = groupLocalService;
-	}
-
-	public void setResourceActionLocalService(
-		ResourceActionLocalService resourceActionLocalService) {
-
-		this.resourceActionLocalService = resourceActionLocalService;
-	}
-
-	public void setResourcePermissionLocalService(
-		ResourcePermissionLocalService resourcePermissionLocalService) {
-
-		this.resourcePermissionLocalService = resourcePermissionLocalService;
-	}
-
-	public void setRoleLocalService(RoleLocalService roleLocalService) {
-		this.roleLocalService = roleLocalService;
-	}
-
-	public void setSortParserProvider(SortParserProvider sortParserProvider) {
-		this.sortParserProvider = sortParserProvider;
-	}
-
-	protected String getApplicationPath() {
-		return "ai-hub";
-	}
-
-	protected Map<String, String> addAction(
-		String actionName,
-		com.liferay.portal.kernel.model.GroupedModel groupedModel,
-		String methodName) {
-
-		return ActionUtil.addAction(
-			actionName, getClass(), groupedModel, methodName,
-			contextScopeChecker, contextUriInfo);
-	}
-
-	protected Map<String, String> addAction(
-		String actionName, Long id, String methodName, Long ownerId,
-		String permissionName, Long siteId) {
-
-		return ActionUtil.addAction(
-			actionName, getClass(), id, methodName, contextScopeChecker,
-			ownerId, permissionName, siteId, contextUriInfo);
-	}
-
-	protected Map<String, String> addAction(
-		String actionName, Long id, String methodName,
-		ModelResourcePermission modelResourcePermission) {
-
-		return ActionUtil.addAction(
-			actionName, getClass(), id, methodName, contextScopeChecker,
-			modelResourcePermission, contextUriInfo);
-	}
-
-	protected Map<String, String> addAction(
-		String actionName, String methodName, String permissionName,
-		Long siteId) {
-
-		return addAction(
-			actionName, siteId, methodName, null, permissionName, siteId);
-	}
-
-	protected <T, R, E extends Throwable> List<R> transform(
-		Collection<T> collection, UnsafeFunction<T, R, E> unsafeFunction) {
-
-		return TransformUtil.transform(collection, unsafeFunction);
-	}
-
-	public static <R, E extends Throwable> R[] transform(
-		int[] array, UnsafeFunction<Integer, R, E> unsafeFunction,
-		Class<? extends R> clazz) {
-
-		return TransformUtil.transform(array, unsafeFunction, clazz);
-	}
-
-	public static <R, E extends Throwable> R[] transform(
-		long[] array, UnsafeFunction<Long, R, E> unsafeFunction,
-		Class<? extends R> clazz) {
-
-		return TransformUtil.transform(array, unsafeFunction, clazz);
-	}
-
-	protected <T, R, E extends Throwable> R[] transform(
-		T[] array, UnsafeFunction<T, R, E> unsafeFunction,
-		Class<? extends R> clazz) {
-
-		return TransformUtil.transform(array, unsafeFunction, clazz);
-	}
-
-	protected <T, R, E extends Throwable> R[] transformToArray(
-		Collection<T> collection, UnsafeFunction<T, R, E> unsafeFunction,
-		Class<? extends R> clazz) {
-
-		return TransformUtil.transformToArray(
-			collection, unsafeFunction, clazz);
-	}
-
-	public static <T, E extends Throwable> boolean[] transformToBooleanArray(
-		Collection<T> collection,
-		UnsafeFunction<T, Boolean, E> unsafeFunction) {
-
-		return TransformUtil.transformToBooleanArray(
-			collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> boolean[] transformToBooleanArray(
-		T[] array, UnsafeFunction<T, Boolean, E> unsafeFunction) {
-
-		return TransformUtil.transformToBooleanArray(array, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> byte[] transformToByteArray(
-		Collection<T> collection, UnsafeFunction<T, Byte, E> unsafeFunction) {
-
-		return TransformUtil.transformToByteArray(collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> byte[] transformToByteArray(
-		T[] array, UnsafeFunction<T, Byte, E> unsafeFunction) {
-
-		return TransformUtil.transformToByteArray(array, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> double[] transformToDoubleArray(
-		Collection<T> collection, UnsafeFunction<T, Double, E> unsafeFunction) {
-
-		return TransformUtil.transformToDoubleArray(collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> double[] transformToDoubleArray(
-		T[] array, UnsafeFunction<T, Double, E> unsafeFunction) {
-
-		return TransformUtil.transformToDoubleArray(array, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> float[] transformToFloatArray(
-		Collection<T> collection, UnsafeFunction<T, Float, E> unsafeFunction) {
-
-		return TransformUtil.transformToFloatArray(collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> float[] transformToFloatArray(
-		T[] array, UnsafeFunction<T, Float, E> unsafeFunction) {
-
-		return TransformUtil.transformToFloatArray(array, unsafeFunction);
-	}
-
-	public static <T, R, E extends Throwable> int[] transformToIntArray(
-		Collection<T> collection, UnsafeFunction<T, R, E> unsafeFunction) {
-
-		return TransformUtil.transformToIntArray(collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> int[] transformToIntArray(
-		T[] array, UnsafeFunction<T, Integer, E> unsafeFunction) {
-
-		return TransformUtil.transformToIntArray(array, unsafeFunction);
-	}
-
-	public static <R, E extends Throwable> List<R> transformToList(
-		int[] array, UnsafeFunction<Integer, R, E> unsafeFunction) {
-
-		return TransformUtil.transformToList(array, unsafeFunction);
-	}
-
-	public static <R, E extends Throwable> List<R> transformToList(
-		long[] array, UnsafeFunction<Long, R, E> unsafeFunction) {
-
-		return TransformUtil.transformToList(array, unsafeFunction);
-	}
-
-	protected <T, R, E extends Throwable> List<R> transformToList(
-		T[] array, UnsafeFunction<T, R, E> unsafeFunction) {
-
-		return TransformUtil.transformToList(array, unsafeFunction);
-	}
-
-	protected <T, R, E extends Throwable> long[] transformToLongArray(
-		Collection<T> collection, UnsafeFunction<T, R, E> unsafeFunction) {
-
-		return TransformUtil.transformToLongArray(collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> long[] transformToLongArray(
-		T[] array, UnsafeFunction<T, Long, E> unsafeFunction) {
-
-		return TransformUtil.transformToLongArray(array, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> short[] transformToShortArray(
-		Collection<T> collection, UnsafeFunction<T, Short, E> unsafeFunction) {
-
-		return TransformUtil.transformToShortArray(collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> short[] transformToShortArray(
-		T[] array, UnsafeFunction<T, Short, E> unsafeFunction) {
-
-		return TransformUtil.transformToShortArray(array, unsafeFunction);
-	}
-
-	protected <T, R, E extends Throwable> List<R> unsafeTransform(
-			Collection<T> collection, UnsafeFunction<T, R, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransform(collection, unsafeFunction);
-	}
-
-	public static <R, E extends Throwable> R[] unsafeTransform(
-			int[] array, UnsafeFunction<Integer, R, E> unsafeFunction,
-			Class<? extends R> clazz)
-		throws E {
-
-		return TransformUtil.unsafeTransform(array, unsafeFunction, clazz);
-	}
-
-	public static <R, E extends Throwable> R[] unsafeTransform(
-			long[] array, UnsafeFunction<Long, R, E> unsafeFunction,
-			Class<? extends R> clazz)
-		throws E {
-
-		return TransformUtil.unsafeTransform(array, unsafeFunction, clazz);
-	}
-
-	protected <T, R, E extends Throwable> R[] unsafeTransform(
-			T[] array, UnsafeFunction<T, R, E> unsafeFunction,
-			Class<? extends R> clazz)
-		throws E {
-
-		return TransformUtil.unsafeTransform(array, unsafeFunction, clazz);
-	}
-
-	protected <T, R, E extends Throwable> R[] unsafeTransformToArray(
-			Collection<T> collection, UnsafeFunction<T, R, E> unsafeFunction,
-			Class<? extends R> clazz)
-		throws E {
-
-		return TransformUtil.unsafeTransformToArray(
-			collection, unsafeFunction, clazz);
-	}
-
-	public static <T, E extends Throwable> boolean[]
-			unsafeTransformToBooleanArray(
-				Collection<T> collection,
-				UnsafeFunction<T, Boolean, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToBooleanArray(
-			collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> boolean[]
-			unsafeTransformToBooleanArray(
-				T[] array, UnsafeFunction<T, Boolean, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToBooleanArray(
-			array, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> byte[] unsafeTransformToByteArray(
-			Collection<T> collection, UnsafeFunction<T, Byte, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToByteArray(
-			collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> byte[] unsafeTransformToByteArray(
-			T[] array, UnsafeFunction<T, Byte, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToByteArray(array, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> double[]
-			unsafeTransformToDoubleArray(
-				Collection<T> collection,
-				UnsafeFunction<T, Double, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToDoubleArray(
-			collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> double[]
-			unsafeTransformToDoubleArray(
-				T[] array, UnsafeFunction<T, Double, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToDoubleArray(
-			array, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> float[] unsafeTransformToFloatArray(
-			Collection<T> collection,
-			UnsafeFunction<T, Float, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToFloatArray(
-			collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> float[] unsafeTransformToFloatArray(
-			T[] array, UnsafeFunction<T, Float, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToFloatArray(array, unsafeFunction);
-	}
-
-	public static <T, R, E extends Throwable> int[] unsafeTransformToIntArray(
-			Collection<T> collection, UnsafeFunction<T, R, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToIntArray(
-			collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> int[] unsafeTransformToIntArray(
-			T[] array, UnsafeFunction<T, Integer, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToIntArray(array, unsafeFunction);
-	}
-
-	public static <R, E extends Throwable> List<R> unsafeTransformToList(
-			int[] array, UnsafeFunction<Integer, R, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToList(array, unsafeFunction);
-	}
-
-	public static <R, E extends Throwable> List<R> unsafeTransformToList(
-			long[] array, UnsafeFunction<Long, R, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToList(array, unsafeFunction);
-	}
-
-	protected <T, R, E extends Throwable> List<R> unsafeTransformToList(
-			T[] array, UnsafeFunction<T, R, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToList(array, unsafeFunction);
-	}
-
-	protected <T, R, E extends Throwable> long[] unsafeTransformToLongArray(
-			Collection<T> collection, UnsafeFunction<T, R, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToLongArray(
-			collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> long[] unsafeTransformToLongArray(
-			T[] array, UnsafeFunction<T, Long, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToLongArray(array, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> short[] unsafeTransformToShortArray(
-			Collection<T> collection,
-			UnsafeFunction<T, Short, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToShortArray(
-			collection, unsafeFunction);
-	}
-
-	public static <T, E extends Throwable> short[] unsafeTransformToShortArray(
-			T[] array, UnsafeFunction<T, Short, E> unsafeFunction)
-		throws E {
-
-		return TransformUtil.unsafeTransformToShortArray(array, unsafeFunction);
-	}
-
-	protected AcceptLanguage contextAcceptLanguage;
-	protected com.liferay.portal.kernel.model.Company contextCompany;
-	protected HttpServletRequest contextHttpServletRequest;
-	protected HttpServletResponse contextHttpServletResponse;
-	protected Object contextScopeChecker;
-	protected UriInfo contextUriInfo;
-	protected com.liferay.portal.kernel.model.User contextUser;
-	protected ExpressionConvert<com.liferay.portal.kernel.search.filter.Filter>
-		expressionConvert;
-	protected FilterParserProvider filterParserProvider;
-	protected GroupLocalService groupLocalService;
-	protected ResourceActionLocalService resourceActionLocalService;
-	protected ResourcePermissionLocalService resourcePermissionLocalService;
-	protected RoleLocalService roleLocalService;
-	protected SortParserProvider sortParserProvider;
-
-	private static final com.liferay.portal.kernel.log.Log _log =
-		LogFactoryUtil.getLog(BaseChatResourceImpl.class);
-
-}
